chore: migrate `<DesignPageNav />`
